third_party/blink/web_tests/external/wpt/html/webappapis/scripting/events/event-handler-attributes-body-alt-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] shadowed pagereveal (document.createElement("body"))
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pageswap (document.createElement("body"))
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed unload (document.createElement("body"))
  assert_equals: body should reflect expected (function) function "() => 0" but got (object) null
[FAIL] not shadowed copy (document.createElement("body"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed cut (document.createElement("body"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed paste (document.createElement("body"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] shadowed pagereveal removal (document.createElement("body"))
  assert_equals: alternative body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed pageswap removal (document.createElement("body"))
  assert_equals: alternative body should reflect changes to itself expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
